refactor(components): 개별 components 하위의 utils 디렉토리를 삭제하고 1depth로 변경'
