+++ src/components/Avatars/AvatarStyle.ts
